Configure isolation level as a parameter.
It used to set the isolation level AUTOCOMMIT everywhere. This isolation level unfortunately is irrelavant for most databases.
The isolation level is now optional, which means if not configured, the SQL engine will apply the default isolation of your backend database. You can now set the isolation level of your choice however  when running a project. Also added a new parameter to CLI to reflect this new parameter.